Fix FP8 MoE CUDA graph benchmark
